Move version under pkg
